app/requirements.txt
fastapi
uvicorn
elasticsearch[async]
python-dotenv
pytz